Rocket landing unit test
Solves 3 Problems

One is a verification of Acikmese Paper of a Spacecraft landing with fix final time.

Another one an enhaced problem of the first, with detailed environment models and using successive convexification

The later one attempts a reprduction of Szmuk paper with drag and using thrust as the control. This later one only works with thrust and the trapezoidal numerical discretization. Requires checks fix the exact dynamics and the acceleration control option